Restart from the beginning, with the new specification
