+++ ra-optimisation/search-tree-rao/src/test/java/com.farao_community.farao.search_tree_rao/config/JsonSearchTreeRaoParametersTest.java
@@ -26,8 +26,6 @@ public class JsonSearchTreeRaoParametersTest extends AbstractConverterTest {
         RaoParameters parameters = new RaoParameters();
         parameters.addExtension(SearchTreeRaoParameters.class, new SearchTreeRaoParameters());
         parameters.getExtension(SearchTreeRaoParameters.class).setRangeActionRao("myRangeActionRao");
-        parameters.getExtension(SearchTreeRaoParameters.class).setDcMode(true);
-        parameters.getExtension(SearchTreeRaoParameters.class).setAcToDcFallback(true);
         roundTripTest(parameters, JsonRaoParameters::write, JsonRaoParameters::read, "/SearchTreeRaoParameters.json");
     }
